Scuicide_game
This is just a pygame game in wihich you play as a spaceship and shoot in to asteroids so they dont collide with you. The idea was that the game deletes itself when you lose so you only have one try bur you know(you can always download the game again so you have as many tryes as you want).